Update refs to refactor AES encryption class.
