+++ phibes/crypto/__init__.py
@@ -5,7 +5,7 @@
 # from phibes.crypto.factory import CryptFactory
 from phibes.crypto.factory import create_crypt, get_crypt, list_crypts
 from phibes.crypto.factory import register_crypt
-from phibes.crypto.crypt_aes256ctr_sha256 import CryptAesCtrPbkdf2Sha256
+from phibes.crypto.crypt_aes_ctr_sha256 import CryptAesCtrPbkdf2Sha256
 
 
 register_crypt(CryptAesCtrPbkdf2Sha256, {'key_rounds': 100100})
